refactor & begin badge
